Revert "build: Add model hack to bypass SafetyNet hardware attestation"
This reverts commit 119fa178ac51760d90ecbad581d8de4631257def.

This breaks a lot of apps because the model name automatically gets
added to the default HTTP User-Agent, and okhttp refuses to send
anything other than printable ASCII characters in headers. We have
another solution coming that will still be handled by the build system,
but it won't be modifying the model prop.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Change-Id: Ib8221722cba9d44af6a9c572eb08b66a08f41b54